+++ 052_add_unaccent_tsv.sql
@@ -10,7 +10,8 @@ AS $$ SELECT ext.unaccent($1); $$;
 
 -- Simple text search config: unaccent filter + simple dictionary
 CREATE TEXT SEARCH DICTIONARY czech_unaccent (
-    Template = unaccent
+    Template = unaccent,
+    Rules = 'unaccent'
 );
 
 CREATE TEXT SEARCH CONFIGURATION czech_simple_unaccent (COPY = simple);