+++ search.py
@@ -17,7 +17,7 @@ RESET = "\033[0m"
 
 DB_PARAMS = {
     "host": "localhost",
-    "port": 5416,
+    "port": 5417,
     "dbname": "hunspell_test",
     "user": "postgres",
     "password": "Password3000!!",